Wrap in class; Support dns lookup;
