+++ scaler/object_storage/defs.h
@@ -10,8 +10,7 @@
 namespace scaler {
 namespace object_storage {
 
-using object_id_t = std::array<uint64_t, 4>;
-// using object_name_t   = std::vector<unsigned char>;
+using object_id_t     = std::array<uint64_t, 4>;
 using object_t        = std::vector<unsigned char>;
 using shared_object_t = std::shared_ptr<object_t>;
 using payload_t       = std::vector<unsigned char>;
@@ -28,19 +27,6 @@ struct ObjectResponseHeader {
     uint64_t payload_length;
     ::ObjectResponseHeader::ObjectResponseType resp_type;
     ObjectResponseHeader(): object_id {}, payload_length {}, resp_type {} {}
-};
-
-// TODO: move only
-struct meta {
-    boost::asio::ip::tcp::socket socket;
-    ObjectRequestHeader request_header;
-    ObjectResponseHeader response_header;
-};
-
-struct object_with_meta {
-    shared_object_t object;
-    std::vector<meta> meta_info;
-    // std::optional<meta> meta_info;
 };
 
 };  // namespace object_storage
